Limor

+++ engine/src/course/java/sdm/engine/engine/Offer.java
@@ -6,13 +6,13 @@ public class Offer {
     private final Item item;
     private final double quantity;
     private final int additionalPrice;
-    private final double totalCost;
+    private final float totalCost;
 
     public Offer(Item item, double quantity, int additionalPrice) {
         this.item = item;
         this.quantity = quantity;
         this.additionalPrice = additionalPrice;
-        this.totalCost = quantity * additionalPrice;
+        this.totalCost = (float) (quantity * additionalPrice);
     }
 
     public Offer(SDMOffer sdmOffer, Item item) {
@@ -31,7 +31,7 @@ public class Offer {
         return additionalPrice;
     }
 
-    public double getTotalCost() {
+    public float getTotalCost() {
         return totalCost;
     }
